requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
